Move App into container directory
